requirements.txt
flask
pygame
pyqt6
requests
flask_wtf
SQLAlchemy-serializer
